--- components/a/Profile/index.tsx
@@ -1,25 +1,0 @@
-import styled from 'styled-components'
-
-export const Inner = styled.div`
-  display: flex;
-  position: absolute;
-  bottom: -30%;
-  left: 2rem;
-`
-
-export const Img = styled.div`
-  width: 5rem;
-  height: 5rem;
-  border-radius: 50%;
-  border: 0.063rem solid #ccc;
-
-  img {
-    border-radius: 50%;
-  }
-`
-
-export const Name = styled.p`
-  padding-top: 1rem;
-  margin-left: 0.5rem;
-  color: #fff;
-`
