+++ components/a/ProfileImg/index.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+export const Img = styled.div`
+  width: 5rem;
+  height: 5rem;
+  border-radius: 50%;
+  border: 0.063rem solid #ccc;
+
+  img {
+    border-radius: 50%;
+  }
+`
