+++ components/a/ProfileName/index.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+export const Name = styled.p`
+  padding-top: 1rem;
+  margin-left: 0.5rem;
+  color: #fff;
+`
