+++ components/a/ProfileWrapper/index.tsx
@@ -1,0 +1,8 @@
+import styled from 'styled-components'
+
+export const Inner = styled.div`
+  display: flex;
+  position: absolute;
+  bottom: -30%;
+  left: 2rem;
+`
