+++ components/m/FixedButton/index.tsx
@@ -1,0 +1,12 @@
+import styled from 'styled-components'
+
+export const Fixed = styled.div`
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  padding: 1rem;
+  box-sizing: border-box;
+  text-align: right;
+  border-top: 1px solid #ccc;
+`
