+++ components/p/Signin/index.stories.tsx
@@ -3,13 +3,13 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { storiesOf } from '@storybook/react'
 import { internet, random } from 'faker'
 import React from 'react'
+import { Button } from '../../a/Button'
+import { Input } from '../../a/Input'
 import {
   Inner as SigninWrapperInner,
   Input as SigninWrapperInput,
   Wrapper as SigninWrapper
-} from '.'
-import { Button } from '../Button'
-import { Input } from '../Input'
+} from '../../a/SigninWrapper'
 
 storiesOf('Signin', module).add('default', () => (
   <SigninWrapper>
